cleaned up components, moved results into new component
